+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Master Dev Branch
 
 ### 2025-10-27
+- Remove WEB_BASE from api calls to remote MAIL API
 - Rename WEB_API to MAIL_API
 
 ### 2025-10-23